Working example and test
made a toolchain test to help prove out the build.  Fixed timing issue (need to count vblanks)

opencode/ollama/gemma4:31b-cloud
opencode/pickle

+++ arduboy-lynx/src/Arduboy2.h
@@ -1516,8 +1516,8 @@ class Arduboy2Base : public Arduboy2Core
 
   // For frame functions
   static uint8_t eachFrameMillis;
-  static uint8_t thisFrameStart;
-  static uint8_t lastFrameDurationMs;
+  static unsigned long thisFrameStart;
+  static unsigned long lastFrameDurationMs;
   static bool justRendered;
 
   // ----- Map of EEPROM addresses for system use-----
